# swagger.yaml
openapi: 3.1.0
info:
  title: KIT KP Monitor API
  description: "Backend API untuk aplikasi Monitoring KP \u2014 KIT"
  version: 1.0.0
paths:
...
  /api/v1/users/me:
    get:
      tags:
      - Users
      summary: Get Me
      operationId: get_me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Get Me Api V1 Users Me Get
      security:
      - HTTPBearer: []
    patch:
      tags:
      - Users
      summary: Update Me
      description: 'Memperbarui data profil (Nama, Divisi, Universitas, Jabatan).

        Data yang tidak ingin diubah dikosongkan (null/omitted).'
      operationId: update_me_api_v1_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
        required: true
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Update Me Api V1 Users Me Patch
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/users/me/avatar:
    post:
      tags:
      - Users
      summary: Update My Avatar
      description: Mengunggah dan memperbarui foto profil (avatar) pengguna.
      operationId: update_my_avatar_api_v1_users_me_avatar_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_update_my_avatar_api_v1_users_me_avatar_post'
        required: true
      responses:
        '200':
          description: Avatar updated successfully
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Update My Avatar Api V1 Users Me Avatar Post
        '400':
          description: No file provided
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/tasks:
    get:
      tags:
      - Tasks
      summary: Get Tasks
      description: 'Mendapatkan daftar tugas harian milik mahasiswa.

        Filter status dan tanggal H-3 s.d. H+3.'
      operationId: get_tasks_api_v1_tasks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          description: 'Filter status: all, done, todo, inprogress'
          default: all
          title: Status
        description: 'Filter status: all, done, todo, inprogress'
      - name: date_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Rentang tanggal H-3 sampai H+3, opsional
          title: Date Filter
        description: Rentang tanggal H-3 sampai H+3, opsional
      responses:
        '200':
          description: Berhasil mendapatkan daftar tugas
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Tasks Api V1 Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Tasks
      summary: Create Task
      description: 'Membuat tugas baru.

        Mendukung upload foto (opsional) menggunakan form-data.'
      operationId: create_task_api_v1_tasks_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_task_api_v1_tasks_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Task Api V1 Tasks Post
        '201':
          description: Berhasil membuat tugas
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}:
    put:
      tags:
      - Tasks
      summary: Update Task
      description: Memperbarui informasi tugas (Judul, deskripsi, status, dll).
      operationId: update_task_api_v1_tasks__task_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Task Api V1 Tasks  Task Id  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Tasks
      summary: Delete Task
      description: Menghapus tugas harian (soft delete).
      operationId: delete_task_api_v1_tasks__task_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Task Api V1 Tasks  Task Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/progress:
    post:
      tags:
      - Tasks
      summary: Add Progress Note
      description: Menambahkan catatan progres (progress note) ke tugas harian tertentu.
      operationId: add_progress_note_api_v1_tasks__task_id__progress_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskProgressNoteCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Add Progress Note Api V1 Tasks  Task Id  Progress
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/photos:
    post:
      tags:
      - Tasks
      summary: Add Task Photo
      description: Menambahkan lampiran foto ke tugas harian.
      operationId: add_task_photo_api_v1_tasks__task_id__photos_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_add_task_photo_api_v1_tasks__task_id__photos_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Add Task Photo Api V1 Tasks  Task Id  Photos Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_add_task_photo_api_v1_tasks__task_id__photos_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_add_task_photo_api_v1_tasks__task_id__photos_post
...
    Body_create_task_api_v1_tasks_post:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        task_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Task Date
        file:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: File
      type: object
      required:
      - title
      - description
      - status
      title: Body_create_task_api_v1_tasks_post
...
    Body_update_my_avatar_api_v1_users_me_avatar_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_update_my_avatar_api_v1_users_me_avatar_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskProgressNoteCreate:
      properties:
        note:
          type: string
          title: Note
      type: object
      required:
      - note
      title: TaskProgressNoteCreate
    TaskUpdateRequest:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        task_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Task Date
      type: object
      title: TaskUpdateRequest
    UserProfileUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        divisi:
          anyOf:
          - type: string
          - type: 'null'
          title: Divisi
        universitas:
          anyOf:
          - type: string
          - type: 'null'
          title: Universitas
        jabatan:
          anyOf:
          - type: string
          - type: 'null'
          title: Jabatan
      type: object
      title: UserProfileUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer